+++ extension/src/popup/constants/routes.ts
@@ -3,9 +3,9 @@ export enum ROUTES {
   welcome = "/",
   account = "/account",
   viewPublicKey = "/account/view-public-key",
+  importAccount = "/account/import",
   accountHistory = "/account-history",
   addAccount = "/add-account",
-  importAccount = "/import-account",
   signTransaction = "/sign-transaction",
   grantAccess = "/grant-access",
   mnemonicPhrase = "/mnemonic-phrase",
