+++ extension/src/popup/views/AddAccount/ImportAccount/styles.scss
@@ -1,0 +1,14 @@
+.ImportAccount {
+  &__warning-header {
+    text-transform: uppercase;
+    font-size: 1rem;
+    line-height: 1.375rem;
+    font-weight: var(--font-weight-medium);
+    margin-bottom: 0.75rem;
+  }
+
+  &__warning_copy {
+    font-size: 0.875rem;
+    line-height: 1.375rem;
+  }
+}
